Add dummy tests

+++ test/conftest.py
@@ -5,4 +5,4 @@ Setup for pytest.
 import numpy as np
 import pytest
 
-np.set_printoptions(precision=10)
+np.set_printoptions(precision=16)
